+++ resources/assets/js/vehicle.js
@@ -3,8 +3,6 @@
  */
 $(document).ready(function () {
 
-    var vehicles_url = "/vehicles";
-
     $('#registration_number').keyup(function(){
         this.value = this.value.toUpperCase();
     });
@@ -13,13 +11,13 @@ $(document).ready(function () {
 
     //display modal form for sending locate request message to vehicle
     function setup_locate_vehicle() {
-        $('.open-modal-location').click(function () {
-
+        $('.open-modal-location').click(function (e) {
+            e.preventDefault();
             var vehicle_id = $('.vehicle_line.selected').attr('data');
             if (vehicle_id) {
-                $('#btn-save-location').val("send");
-                $('#location_vehicle_id').val(vehicle_id);
-                $('#locationModal').modal('show');
+                $('#btn-save-locateVehicle').val("send");
+                $('#locateVehicle_id').val(vehicle_id);
+                $('#locateVehicleModal').modal('show');
             }
         });
     }
@@ -27,23 +25,26 @@ $(document).ready(function () {
 
     //display modal form for vehicle editing
     function setup_edit_vehicle() {
-        $('.open-modal-vehicle').click(function () {
-            var vehicle_id = $(this).val();
+        $('.open-modal-vehicle').click(function (e) {
+            e.preventDefault();
+            var vehicle_id = $('.vehicle_line.selected').attr('data');
+            if (vehicle_id) {
 
-            $.get(vehicles_url + '/' + vehicle_id, function (data) {
-                //success data
-                console.log(data);
-                $('#vehicle_id').val(data._id);
-                $('#registration_number').val(data.registration_number);
-                $('#vehicle_mobile_phone_number').val(data.mobile_phone_number);
-                $('#tracker_imei_number').val(data.tracker_imei_number);
-                $('#btn-save-vehicle').val("update");
-                $('#vehicleModal').modal('show');
-            }).fail(function(data){
-                var newDoc = document.open("text/html", "replace");
-                newDoc.write(data.responseText);
-                newDoc.close();
-            });
+                $.get( '/vehicles/' + vehicle_id, function (data) {
+                    //success data
+                    console.log(data);
+                    $('#vehicle_id').val(data._id);
+                    $('#registration_number').val(data.registration_number);
+                    $('#vehicle_mobile_phone_number').val(data.mobile_phone_number);
+                    $('#tracker_imei_number').val(data.tracker_imei_number);
+                    $('#btn-save-vehicle').val("update");
+                    $('#vehicleModal').modal('show');
+                }).fail(function(data){
+                    var newDoc = document.open("text/html", "replace");
+                    newDoc.write(data.responseText);
+                    newDoc.close();
+                });
+            }
         });
     }
 
@@ -61,45 +62,54 @@ $(document).ready(function () {
     //delete vehicle and remove it from list
 
     function setup_delete_vehicle() {
-        $('.delete-vehicle').click(function () {
-            $.ajaxSetup({
-                headers: {
-                    'X-CSRF-TOKEN': $('meta[name="_token"]').attr('content')
-                }
-            });
+        $('.delete-vehicle').click(function (e) {
+            e.preventDefault();
+            var vehicle_id = $('.vehicle_line.selected').attr('data');
+            if (vehicle_id) {
+                $.ajaxSetup({
+                    headers: {
+                        'X-CSRF-TOKEN': $('meta[name="_token"]').attr('content')
+                    }
+                });
+                $.ajax({
 
-            var vehicle_id = $(this).val();
+                    type: "DELETE",
+                    url: '/vehicles/' + vehicle_id,
+                    success: function (data) {
+                        console.log(data);
 
-            $.ajax({
-
-                type: "DELETE",
-                url: vehicles_url + '/' + vehicle_id,
-                success: function (data) {
-                    console.log(data);
-
-                    $("#vehicle" + vehicle_id).remove();
-                },
-                error: function (data) {
-                    console.log('Error:', data);
-                    var newDoc = document.open("text/html", "replace");
-                    newDoc.write(data.responseText);
-                    newDoc.close();
-                }
-            });
+                        $("#vehicle" + vehicle_id).remove();
+                    },
+                    error: function (data) {
+                        console.log('Error:', data);
+                        var newDoc = document.open("text/html", "replace");
+                        newDoc.write(data.responseText);
+                        newDoc.close();
+                    }
+                });
+            }
         });
     }
 
     setup_delete_vehicle();
 
+    function setup_reset_vehicle_form(){
+        $('#vehicleForm').on("reset",function(){
+            var $form = $('#vehicleForm');
+            $form.find('.help-block').remove();
+            $form.find('.form-group').removeClass('has-error');
+        });
+    }
+    setup_reset_vehicle_form();
+
     //create new vehicle / update existing vehicle
     $("#btn-save-vehicle").click(function (e) {
+        e.preventDefault();
         $.ajaxSetup({
             headers: {
                 'X-CSRF-TOKEN': $('meta[name="_token"]').attr('content')
             }
         });
-
-        e.preventDefault();
 
         var formData = {
             registration_number: $('#registration_number').val(),
@@ -112,11 +122,11 @@ $(document).ready(function () {
 
         var type = "POST"; //for creating new resource
         var vehicle_id = $('#vehicle_id').val();
-        var my_vehicle_url = vehicles_url;
+        var vehicle_url = '/vehicles';
 
         if (state == "update") {
             type = "PUT"; //for updating existing resource
-            my_vehicle_url += '/' + vehicle_id;
+            vehicle_url += '/' + vehicle_id;
         }
 
         console.log(formData);
@@ -124,7 +134,7 @@ $(document).ready(function () {
         $.ajax({
 
             type: type,
-            url: my_vehicle_url,
+            url: vehicle_url,
             data: formData,
             dataType: 'json',
             success: function (data) {
@@ -136,11 +146,8 @@ $(document).ready(function () {
                         .clone(false)
                         .appendTo('#vehicle_list')
                         .attr("id", "vehicle" + data._id)
-                        .attr("data",data._id);
-                    $("#vehicle" + data._id + ' button.open-modal-location').val(data._id);
-                    $("#vehicle" + data._id + ' button.open-modal-vehicle').val(data._id);
-                    $("#vehicle" + data._id + ' button.delete-vehicle').val(data._id);
-                    $("#vehicle" + data._id).css('display', '');
+                        .attr("data",data._id)
+                        .show();
                 }
                 $("#vehicle" + data._id + " .registration_number").text(data.registration_number);
 
@@ -148,9 +155,6 @@ $(document).ready(function () {
                 $('#vehicleModal').modal('hide');
 
                 setup_select('vehicle',true);
-                setup_locate_vehicle();
-                setup_edit_vehicle();
-                setup_delete_vehicle();
                 adjust_fluid_columns();
                 setup_sse();
